requirements.txt
coloredlogs==15.0.1
pandas==1.5.3
pika==1.3.1
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.2
